@@ -1,7 +1,9 @@
-import Vue from 'vue';
-import Vuetify from 'vuetify/lib';
+import 'vuetify/styles'
+import { createVuetify } from 'vuetify'
+import * as components from 'vuetify/components'
+import * as directives from 'vuetify/directives'
 
-Vue.use(Vuetify);
-
-export default new Vuetify({
-});
+export default createVuetify({
+  components,
+  directives,
+})
